# api/openapi/stacks.yaml
openapi: 3.0.3
info:
  title: FloatLab Stacks API
  version: 0.1.0
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        maxLength: 255
      description: Unique key used to safely retry this mutation.
  schemas:
    StackState:
      type: string
      enum:
        - Provisioning
        - Idle
        - Starting
        - RunningPrimary
        - FailingOver
        - RunningBackup
        - Restoring
        - Stopping
        - Updating
        - RollingBack
        - Failed
      description: Current lifecycle state of the stack.
...
    ComposeValidation:
      type: object
      required: [compose_file]
      properties:
        name:
          type: string
          description: Required for a new stack; ignored when stack_id is supplied.
        compose_file:
          type: string
        stack_id:
          type: string
          format: uuid
          description: Existing stack to validate an edit for.
    ComposeUpdate:
      type: object
      required: [compose_file]
      properties:
        compose_file:
          type: string
    StackUpgrade:
      type: object
      required: [images]
      properties:
        images:
          type: object
          minProperties: 1
          additionalProperties:
            type: string
          description: Compose service names mapped to replacement image references.
...
    Error:
      type: object
      required: [status, message]
      properties:
        status:
          type: integer
        message:
          type: string
        code:
          type: string
paths:
  /stacks/validate:
    post:
      summary: Validate a stack Compose document without mutating state
      operationId: validateStackCompose
      tags: [stacks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComposeValidation'
      responses:
        '204':
          description: Compose document is valid.
        '400':
          description: Compose validation failure.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Existing stack not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /stacks:
    get:
      summary: List all stacks
      operationId: listStacks
      tags: [stacks]
      responses:
        '200':
          description: List of all stacks.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Stack'
    post:
      summary: Create a stack
      operationId: createStack
      tags: [stacks]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StackCreate'
      responses:
        '201':
          description: Stack created; dataset provisioned; state=Provisioning.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stack'
        '400':
          description: Invalid request body or compose_file validation failure.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Stack name already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /stacks/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get a stack by ID
      operationId: getStack
      tags: [stacks]
      responses:
        '200':
          description: Stack detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stack'
        '404':
          description: Stack not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Delete a stack
      operationId: deleteStack
      tags: [stacks]
      description: Stops the stack and removes active metadata. The project dataset and configuration are retained unless purged.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - name: purge
          in: query
          schema:
            type: boolean
            default: false
          description: Also destroy the ZFS dataset and all snapshots.
      responses:
        '202':
          description: Delete operation accepted.
        '409':
          description: Stack is not in a deletable state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /stacks/{id}/compose:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    put:
      summary: Update the compose file for a stack
      operationId: updateStackCompose
      tags: [stacks]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComposeUpdate'
      responses:
        '200':
          description: Compose file updated. Stack state unchanged.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stack'
        '400':
          description: Compose validation failure.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /stacks/{id}/start:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      summary: Create or update and start a stack
      operationId: startStack
      tags: [stacks]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: false
        content:
          application/yaml:
            schema: {type: string}
          application/x-yaml:
            schema: {type: string}
          application/x-tar:
            schema: {type: string, format: binary}
      responses:
        '202':
          description: Start accepted; state transitions asynchronously.
        '409':
          description: Stack is not in a startable state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /stacks/{id}/stop:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      summary: Stop a stack (RunningPrimary → Stopping → Idle)
      operationId: stopStack
      tags: [stacks]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202':
          description: Stop accepted.
        '409':
          description: Stack is not in a stoppable state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /stacks/{id}/upgrade:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      summary: Upgrade selected service images with automatic rollback
      operationId: upgradeStack
      tags: [stacks]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StackUpgrade'
      responses:
        '202':
          description: Upgrade accepted. The Location header identifies the resumable operation.
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                required: [stack_id, state, operation_id]
                properties:
                  stack_id:
                    type: string
                    format: uuid
                  state:
                    $ref: '#/components/schemas/StackState'
                  operation_id:
                    type: string
                    format: uuid
        '409':
          description: Stack is not running on its primary node.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /stacks/{id}/failover:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      summary: Trigger manual failover (RunningPrimary → FailingOver → RunningBackup)
      operationId: triggerFailover
      tags: [stacks]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202':
          description: Failover accepted.
        '409':
          description: Stack is not in RunningPrimary state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /stacks/{id}/restore:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      summary: Restore primary (RunningBackup → Restoring → RunningPrimary)
      operationId: restoreStack
      tags: [stacks]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202':
          description: Restore accepted.
        '409':
          description: Stack is not in RunningBackup state.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /stacks/{id}/state:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get current state of a stack
      operationId: getStackState
      tags: [stacks]
      responses:
        '200':
          description: Current state.
          content:
            application/json:
              schema:
                type: object
                required: [state]
                properties:
                  state:
                    $ref: '#/components/schemas/StackState'
  /stacks/{id}/containers:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: List containers for a stack (across all nodes)
      operationId: getStackContainers
      tags: [stacks]
      responses:
        '200':
          description: List of containers.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Container'
...
  /stacks/{id}/containers/{containerId}/start:
    parameters:
      - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
      - {name: containerId, in: path, required: true, schema: {type: string}}
    post:
      summary: Start one container in a running stack
      operationId: startStackContainer
      tags: [stacks]
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      responses:
        '200':
          description: Updated container state.
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Container'}
        '404': {description: Stack or container not found.}
        '409': {description: Stack is not running.}
...
  /stacks/{id}/restart:
    parameters:
      - {name: id, in: path, required: true, schema: {type: string, format: uuid}}
    post:
      summary: Restart a running stack
      operationId: restartStack
      tags: [stacks]
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      responses:
        '202': {description: Restart operation accepted.}
        '409': {description: Stack is not running.}
...